(interpreter) changed the signature of compareYUV in DisplayYUV.java to use BigInteger parameters

+++ eclipse/plugins/net.sf.orcc.simulators/src/std/video/impl/DisplayYUV.java
@@ -103,7 +103,7 @@ public class DisplayYUV {
 	 */
 	public static void compareYUV_comparePicture(short[] pictureBufferY,
 			short[] pictureBufferU, short[] pictureBufferV,
-			Integer pictureWidth, Integer pictureHeight) {
+			BigInteger pictureWidth, BigInteger pictureHeight) {
 
 	}
